Fix assertion when starting Python GTK application
Python's gettext doesn't use the underlying gettext implementation[1] in
the glibc and thus will not call textdomain() to set the default domain,
as it doesn't need it.

Remove assertion when textdomain() is not called. The end result in a C
program will be a half-translated program, which should be obvious from
the UI.

Closes: #7

[1]: https://docs.python.org/3/library/gettext.html
> The gettext module defines the following API, which is very similar to the GNU gettext API.